+++ pdpyras.py
@@ -170,6 +170,24 @@ def resource_name(obj_type):
     else:
         return obj_type+'s'
 
+def resource_path(method):
+    """
+    API call decorator that allows passing a resource dict as the path/URL
+
+    Most resources returned by the API will contain a ``self`` attribute that is
+    the URL of the resource itself.
+
+    Using this decorator allows the implementer to pass either a URL/path or
+    such a resource dictionary as the ``path`` argument, thus eliminating the
+    need to re-construct the resource URL or hold it in a temporary variable.
+    """
+    def call(self, resource, **kw):
+        url = resource
+        if type(resource) is dict and 'self' in resource: # passing an object
+            url = resource['self']
+        return method(self, url, **kw)
+    return call
+
 def tokenize_url_path(url, baseurl='https://api.pagerduty.com'):
     """
     Classifies a URL according to some global patterns in the API.
@@ -1154,19 +1172,43 @@ class APISession(PDSession):
         :type suffix: str
         :rtype: str
         """
-        my_suffix = "" if suffix is None else "#"+suffix 
+        my_suffix = "" if suffix is None else "#"+suffix
         path_str = '/'.join(tokenize_url_path(path, baseurl=self.url))
         return '%s:%s'%(method.lower(), path_str)+my_suffix
 
-    def rdelete(self, path, **kw):
-        raise_on_error(self.delete(path, **kw))
+    @resource_path
+    def rdelete(self, resource, **kw):
+        """
+        Delete a resource.
 
+        :param resource:
+            The path/URL to which to send the request, or a dict object
+            representing an API resource that contains an item with key ``self``
+            whose value is the URL of the resource.
+        :param \*\*kw:
+            Keyword arguments to pass to ``requests.Session.delete``
+        :type path: str or dict
+        """
+        raise_on_error(self.delete(resource, **kw))
+
+    @resource_path
     @resource_envelope
-    def rget(self, path, **kw):
+    def rget(self, resource, **kw):
         """
         Retrieve a resource and return the encapsulated object in the response
+
+        :param resource:
+            The path/URL to which to send the request, or a dict object
+            representing an API resource that contains an item with key ``self``
+            whose value is the URL of the resource.
+        :param \*\*kw:
+            Keyword arguments to pass to ``requests.Session.rget``
+        :returns:
+            Dictionary representation of the object.
+        :type resource: str or dict
+        :rtype dict:
         """
-        return self.get(path, **kw)
+        return self.get(resource, **kw)
 
     @resource_envelope
     def rpost(self, path, **kw):
@@ -1176,22 +1218,34 @@ class APISession(PDSession):
         Returns the dictionary object representation if creating it was
         successful.
 
-        :param path: The path/URL to which to send the POST request.
-        :param \*\*kw: Keyword arguments to pass to ``requests.Session.post``
-        :returns: Dictionary representation of the created object
-        :rtype: 
+        :param path:
+            The path/URL to which to send the POST request, which should be an
+            index endpoint.
+        :param \*\*kw:
+            Keyword arguments to pass to ``requests.Session.post``
+        :returns:
+            Dictionary representation of the created object
+        :type path: str
+        :rtype dict:
         """
         return self.post(path, **kw)
 
+    @resource_path
     @resource_envelope
-    def rput(self, path, **kw):
+    def rput(self, resource, **kw):
         """
         Update an individual resource, returning the encapsulated object.
 
-        :param path: The path/URL to which to send the PUT request.
-        :param \*\*kw: Keyword arguments to pass to ``requests.Session.put``
+        :param resource:
+            The path/URL to which to send the request, or a dict object
+            representing an API resource that contains an item with key ``self``
+            whose value is the URL of the resource.
+        :param \*\*kw:
+            Keyword arguments to pass to ``requests.Session.put``
+        :returns:
+            Dictionary representation of the updated object
         """
-        return self.put(path, **kw)
+        return self.put(resource, **kw)
 
     @property
     def subdomain(self):